Use the upstream oai-pmh-client transient retries and drop ResilientOAIClient
oai-pmh-client 1.1.0 (wellcomecollection/oai-pmh#13) has the resilience
behaviour natively: transient failures (empty bodies, transport errors, 5xx)
are retried with exponential backoff on initial requests, and failed
resumption token requests raise ResumptionTokenFailedError instead of being
replayed. The window harvester's broad per-window exception handling already
restarts the list operation, which is the correct recovery for burned tokens.

The dependency is pinned to the merged upstream commit; it should be repinned
to tag v1.1.0 once the release PR (wellcomecollection/oai-pmh#14) is merged.

The OAI_EMPTY_BODY_RETRIES / FOLIO_OAI_EMPTY_BODY_RETRIES settings are renamed
to OAI_TRANSIENT_RETRIES / FOLIO_OAI_TRANSIENT_RETRIES to match the upstream
parameter; neither env var is set anywhere in terraform, so the rename only
touches defaults. Retry behaviour tests now live upstream; the downstream
tests keep the factory wiring assertions.
