+++ catalogue_graph/pyproject.toml
@@ -40,7 +40,9 @@ dependencies = [
 package = true
 
 [tool.uv.sources]
-oai-pmh-client = { git = "https://github.com/wellcomecollection/oai-pmh.git", tag = "v1.0.1" }
+# Pinned to the merged transient-retries commit; repin to tag = "v1.1.0" once
+# https://github.com/wellcomecollection/oai-pmh/pull/14 is released.
+oai-pmh-client = { git = "https://github.com/wellcomecollection/oai-pmh.git", rev = "cb928e0ce9c5f8cfea34c6a05a1a243951bd549e" }
 
 
 [dependency-groups]
