+++ catalogue_graph/src/adapters/extractors/oai_pmh/axiell/clients.py
@@ -9,7 +9,6 @@ from oai_pmh_client.client import OAIClient
 
 from adapters.extractors.oai_pmh.axiell import config
 from adapters.extractors.oai_pmh.http_client import OAIPMHHTTPClient
-from adapters.extractors.oai_pmh.resilient_client import ResilientOAIClient
 from utils.aws import get_ssm_parameter
 
 # Axiell uses a custom "Token" header for authentication
@@ -42,13 +41,11 @@ def build_http_client(*, token: str | None = None) -> httpx.Client:
 def build_oai_client(*, http_client: httpx.Client | None = None) -> OAIClient:
     """Build an OAI-PMH client for Axiell."""
     client = http_client or build_http_client()
-    return ResilientOAIClient(
+    return OAIClient(
         _oai_endpoint(),
         client=client,
-        empty_body_retries=config.OAI_EMPTY_BODY_RETRIES,
-        empty_body_backoff_factor=config.OAI_BACKOFF_FACTOR,
-        empty_body_backoff_max=config.OAI_BACKOFF_MAX,
         max_request_retries=config.OAI_MAX_RETRIES,
         request_backoff_factor=config.OAI_BACKOFF_FACTOR,
         request_max_backoff=config.OAI_BACKOFF_MAX,
+        max_transient_retries=config.OAI_TRANSIENT_RETRIES,
     )
