+++ catalogue_graph/src/adapters/extractors/oai_pmh/folio/config.py
@@ -75,9 +75,9 @@ OAI_MAX_READ_TIMEOUT = float(os.getenv("FOLIO_OAI_MAX_READ_TIMEOUT", "60.0"))
 OAI_MAX_RETRIES = int(os.getenv("FOLIO_OAI_MAX_RETRIES", "1"))
 OAI_BACKOFF_FACTOR = float(os.getenv("FOLIO_OAI_BACKOFF_FACTOR", "0.75"))
 OAI_BACKOFF_MAX = float(os.getenv("FOLIO_OAI_BACKOFF_MAX", "5.0"))
-# Retries for empty-body (HTTP 200, no content) responses; shares the backoff
-# ladder configured by FOLIO_OAI_BACKOFF_FACTOR / FOLIO_OAI_BACKOFF_MAX.
-OAI_EMPTY_BODY_RETRIES = int(os.getenv("FOLIO_OAI_EMPTY_BODY_RETRIES", "3"))
+# Retries for transient failures (empty bodies, transport errors, 5xx); shares
+# the backoff ladder configured by FOLIO_OAI_BACKOFF_FACTOR / FOLIO_OAI_BACKOFF_MAX.
+OAI_TRANSIENT_RETRIES = int(os.getenv("FOLIO_OAI_TRANSIENT_RETRIES", "3"))
 
 
 # ---------------------------------------------------------------------------
